requirements.txt
Flask==2.3.2
Pillow==9.5.0
